Assimilate
The logger is used rather than println().
Some return values are Arrays instead of Seq.
The code looks for like other nearby code.
The breeze version is lowered to support Scala 2.11.  Later the version number of the dependency will be conditional based on Scala version.  This can be done sooner if need be.